# openapi.yaml
openapi: 3.0.3
info:
  title: Yeelo Homeopathy ERP API
  description: |
    Complete REST API for Yeelo Homeopathy Enterprise Resource Planning System.
    
    ## Features
    - Product & Inventory Management
    - Sales & Purchase Orders
    - Customer & Vendor Management
    - Financial Accounting
    - HR & Payroll
    - AI-Powered Analytics
    - MeiliSearch Integration
    - Multi-branch Support
    
    ## Authentication
    All endpoints require JWT token in Authorization header:
    ```
    Authorization: Bearer <token>
    ```
  version: 2.1.0
  contact:
    name: Yeelo Support
    email: support@yeelo.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3005/api/erp
    description: Development server
  - url: https://api.yeelo.com/api/erp
    description: Production server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /ai/chat:
    post:
      tags: [AI]
      summary: Chat with AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                thread_id:
                  type: string
                assistant_type:
                  type: string
                  enum: [erp, forecast, prescription, marketing]
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  response:
                    type: string
                  thread_id:
                    type: string
...